Final wave of the pre-M6 cleanup ramp.  Closes ~104 wave-6-cleanup
audit IDs + reactive runtime polish bundle (REACT-POLISH-001..009;
-003 deferred) + 5 carry-forwards.  Strict tooling promoted to
hard-fail (cppcheck-strict 135 -> 0; tidy-strict 23 -> 0).
Docstring coverage gate live on every header-exposed symbol in
include/urbi/ and subsystem-public src/<subsys>/u*.h headers.
Footprint within +/-2% of v0.5.6 baseline (Phase 18 cheap-win
tightening).  Bytecode + wire-format byte-identical against
v0.5.7-fixes; only diff is the LEX-034 fixture addition (149 vs
148 fixtures).

Public API: urbi_run_chunk and urbi_run_script now take
const struct UModule *module (was non-const).  Source-compatible for
direct callers; technically breaking for embedders that explicitly
typed forwarding declarations as non-const.

Cleanup ramp COMPLETE.  v0.5 era closes; M6 stdlib next.

Plan: docs/superpowers/plans/2026-05-08-v0.5.8-cleanup.md
Retrospective: docs/milestones/v0.5.x-cleanup.md

+++ src/object/umodule_instance.c
@@ -192,7 +192,7 @@ urbi_module_instance_create(struct UVM *vm, UModule *m)
     /* entries[1..n-1]: parallel to module->nested[].  Each gets its own
      * slice of the trailing IC region; unfilled sites have topology_gen == 0
      * (the sentinel per pre-M4 topology-generation spec §3.1). */
-    for (uint16_t i = 0U; i < m->nested_count; i++) {
+    for (size_t i = 0U; i < m->nested_count; i++) {
         UProto *p = m->nested[i];
         if (p != NULL &&
             !intern_ic_names_from_strs(vm, p->ic_count, &p->ic_names,
@@ -223,10 +223,11 @@ urbi_module_instance_create(struct UVM *vm, UModule *m)
 void
 urbi_module_instance_destroy(struct UVM *vm, UModuleInstance *mi)
 {
-    /* GC-managed; sweep reaps both cells when no roots reach mi.  No
-     * explicit teardown at T16 — IC entries hold no host-owned resources
-     * yet (T22+ may revisit if IC fill installs anything that needs an
-     * explicit finalizer). */
+    /* AUDIT: OBJ-027 — body intentionally empty at v1.0.  Both cells are
+     * GC-managed; sweep reaps them when no roots reach mi.  Symbol kept
+     * because the public-API contract (paired create/destroy) survives
+     * into M7, where this hook may grow semantics (host-visible registry
+     * detach, etc.).  No churn for callers when that lands. */
     (void)vm;
     (void)mi;
 }
